+++ .markdownlint.js
@@ -6,5 +6,5 @@ module.exports = {
   extends: base,
   files: ["content/**/*.mdx"],
   MD033: { allowed_elements: ["Tip", "Warning", "button"] },
-  MD013: false
+  MD013: false,
 };