+++ components/CmdProvider.tsx
@@ -1,0 +1,13 @@
+import { PropsWithChildren } from 'react';
+import { useCmd } from '../hooks/useCmd';
+import { Cmd } from './CmdK';
+
+export const CmdProvider: React.FC<PropsWithChildren> = ({ children }) => {
+  useCmd();
+  return (
+    <>
+      {children}
+      <Cmd />
+    </>
+  );
+};