+++ components/MainLayout.tsx
@@ -1,15 +1,9 @@
-import Head from "next/head";
+import Head from 'next/head';
+import { PropsWithChildren } from 'react';
 
-import Logo from "./Logo";
-import Foot from "./Foot";
-import Nav from "../components/Nav";
+import TopNav from './TopNav';
 
-interface PropTypes {
-  children: React.ReactNode;
-  pages: string[];
-}
-
-const Layout = ({ children }: PropTypes) => {
+const Layout: React.FC<PropsWithChildren> = ({ children }) => {
   return (
     <>
       <Head>
@@ -39,18 +33,8 @@ const Layout = ({ children }: PropTypes) => {
         <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5" />
         <meta name="theme-color" content="#ffffff" />
       </Head>
-      <div className="app-container">
-        <main>
-          <Nav>
-            <section className="flex flex-1 flex-col lg:pl-64">
-              {children}
-            </section>
-          </Nav>
-          <footer>
-            <Foot />
-          </footer>
-        </main>
-      </div>
+      <TopNav />
+      {children}
     </>
   );
 };