+++ components/Tip.tsx
@@ -7,8 +7,8 @@ interface Props {
 
 const Tip = (props: Props) => {
   return (
-    <div className="tip">
-      <div className="tip-content-title">{props.title}</div>
+    <div className="border-l-4 p-5 pl-8 border-l-green-500">
+      <div className="font-bold mb-4 text-lg">{props.title}</div>
       <div className="tip-content">{props.children}</div>
     </div>
   );