+++ components/TopNav.tsx
@@ -1,0 +1,96 @@
+import FutureImage from 'next/future/image';
+import { Disclosure } from '@headlessui/react';
+import { MagnifyingGlassIcon } from '@heroicons/react/20/solid';
+import { Bars3Icon, XMarkIcon } from '@heroicons/react/24/outline';
+import Link from 'next/link';
+
+export default function TopNav() {
+  return (
+    <Disclosure as="nav" className="bg-white shadow sticky lg:top-0 lg:z-10">
+      {({ open: openDisclosure }) => (
+        <>
+          <div className="mx-auto px-2 sm:px-4 lg:px-8">
+            <div className="flex h-16 justify-between">
+              <div className="flex px-2 lg:px-0">
+                <Link href="/home">
+                  <div className="flex flex-shrink-0 items-center cursor-pointer">
+                    <FutureImage
+                      src="/metamask.svg"
+                      alt="Metamask"
+                      width={45}
+                      height={45}
+                    />
+                  </div>
+                </Link>
+                <div className="hidden lg:ml-6 lg:flex lg:space-x-8">
+                  <Link href="/guide/introduction">
+                    <a className="inline-flex items-center border-b-2 border-transparent px-1 pt-1 text-sm font-medium text-gray-500 hover:border-gray-300 hover:text-gray-700">
+                      Guide
+                    </a>
+                  </Link>
+                  <Link href="/mobile">
+                    <a className="inline-flex items-center border-b-2 border-transparent px-1 pt-1 text-sm font-medium text-gray-500 hover:border-gray-300 hover:text-gray-700">
+                      Mobile
+                    </a>
+                  </Link>
+                </div>
+              </div>
+              <div className="flex flex-1 items-center justify-center px-2 lg:ml-6 lg:justify-end">
+                <div className="w-full max-w-lg lg:max-w-xs">
+                  <label htmlFor="search" className="sr-only">
+                    Search
+                  </label>
+                  <div className="relative">
+                    <div className="pointer-events-none absolute inset-y-0 left-0 flex items-center pl-3">
+                      <MagnifyingGlassIcon
+                        className="h-5 w-5 text-gray-400"
+                        aria-hidden="true"
+                      />
+                    </div>
+                    <input
+                      id="search"
+                      name="search"
+                      className="block w-full rounded-md border border-gray-300 bg-white py-2 pl-10 pr-3 leading-5 placeholder-gray-500 focus:border-indigo-500 focus:placeholder-gray-400 focus:outline-none focus:ring-1 focus:ring-indigo-500 sm:text-sm"
+                      placeholder="Search"
+                      type="search"
+                    />
+                  </div>
+                </div>
+              </div>
+              <div className="flex items-center lg:hidden">
+                {/* Mobile menu button */}
+                <Disclosure.Button className="inline-flex items-center justify-center rounded-md p-2 text-gray-400 hover:bg-gray-100 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500">
+                  <span className="sr-only">Open main menu</span>
+                  {openDisclosure ? (
+                    <XMarkIcon className="block h-6 w-6" aria-hidden="true" />
+                  ) : (
+                    <Bars3Icon className="block h-6 w-6" aria-hidden="true" />
+                  )}
+                </Disclosure.Button>
+              </div>
+            </div>
+          </div>
+
+          <Disclosure.Panel className="lg:hidden">
+            <div className="space-y-1 pt-2 pb-3">
+              <Disclosure.Button
+                as="a"
+                href="#"
+                className="block border-l-4 border-transparent py-2 pl-3 pr-4 text-base font-medium text-gray-600 hover:border-gray-300 hover:bg-gray-50 hover:text-gray-800"
+              >
+                Guide
+              </Disclosure.Button>
+              <Disclosure.Button
+                as="a"
+                href="#"
+                className="block border-l-4 border-transparent py-2 pl-3 pr-4 text-base font-medium text-gray-600 hover:border-gray-300 hover:bg-gray-50 hover:text-gray-800"
+              >
+                Mobile
+              </Disclosure.Button>
+            </div>
+          </Disclosure.Panel>
+        </>
+      )}
+    </Disclosure>
+  );
+}