+++ layout/Logo.tsx
@@ -1,5 +1,0 @@
-const Logo = () => {
-  return <div className={`logo`}>MetaMask Docs</div>;
-};
-
-export default Logo;