+++ pages/_document.tsx
@@ -1,6 +1,6 @@
-import { Html, Head, Main, NextScript } from "next/document";
+import { Html, Head, Main, NextScript } from 'next/document';
 
-export default function Document() {
+export default function NextDocument() {
   return (
     <Html className="h-full">
       <Head />