+++ pages/home.tsx
@@ -1,11 +1,7 @@
 import type { NextPage } from 'next';
 
 const Home: NextPage = () => {
-  return (
-    <div>
-      <h1>Home Page</h1>
-    </div>
-  );
+  return <h1>Home Page</h1>;
 };
 
 export default Home;